pyproject.toml
[project]
name = "base-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain>=1.2.14",
    "langchain-deepseek>=1.0.1",
    "langgraph>=1.1.4",
    "python-dotenv>=1.2.2",
]
